+++ cpp/open3d/t/geometry/kernel/VoxelBlockGrid.cpp
@@ -219,7 +219,8 @@ void EstimateRange(const core::Tensor& block_keys,
                    int64_t block_resolution,
                    float voxel_size,
                    float depth_min,
-                   float depth_max) {
+                   float depth_max,
+                   core::Tensor& fragment_buffer) {
     static const core::Device host("CPU:0");
     core::Tensor intrinsics_d = intrinsics.To(host, core::Float64).Contiguous();
     core::Tensor extrinsics_d = extrinsics.To(host, core::Float64).Contiguous();
@@ -227,12 +228,12 @@ void EstimateRange(const core::Tensor& block_keys,
     if (block_keys.IsCPU()) {
         EstimateRangeCPU(block_keys, range_minmax_map, intrinsics_d,
                          extrinsics_d, h, w, down_factor, block_resolution,
-                         voxel_size, depth_min, depth_max);
+                         voxel_size, depth_min, depth_max, fragment_buffer);
     } else if (block_keys.IsCUDA()) {
 #ifdef BUILD_CUDA_MODULE
         EstimateRangeCUDA(block_keys, range_minmax_map, intrinsics_d,
                           extrinsics_d, h, w, down_factor, block_resolution,
-                          voxel_size, depth_min, depth_max);
+                          voxel_size, depth_min, depth_max, fragment_buffer);
 #else
         utility::LogError("Not compiled with CUDA, but CUDA device is used.");
 #endif